Fix version mixup in CompareSimple
'l' was incorrectly representing the value of the right argument
(version2) and 'r' of the left argument (version1) while it should
be exactly the other way.

This lead to odd bugs:

* CompareSimple("1.8.2.0", "1.8.3.0-dev") would return '-1' even though
  the returned value should be '1'.

The test suite was set up in a way to make for the mixup so that's
probably the reason why this was not detected by the author.

This changes the 'l' and 'r' variables, updates return values to be the
same as before (to prevent revriting the test suite) and updates the
documentation.